- Added untested DCC large file support. It should work as before,   but the large file part is untested.
git-svn-id: svn://red.pomac.com/rage@87 55c13d69-e7e2-0310-b639-a4a306287c42

+++ trunk/src/common/dcc.h
@@ -36,10 +36,10 @@ struct DCC
 	unsigned int lastcpspos;
 	int maxcps;
 
-	unsigned int size;
-	unsigned int resumable;
-	unsigned int ack;
-	unsigned int pos;
+	off_t size;
+	off_t resumable;
+	off_t ack;
+	off_t pos;
 	time_t starttime;
 	time_t offertime;
 	time_t lasttime;
